added a close buttton to drawer menu, put the folders back and added mono.0 and mono.1000 to theme file
