+++ app/index.tsx
@@ -1,7 +1,5 @@
 import { Redirect } from 'expo-router';
 
 export default function Index() {
-  return (
-    <Redirect href={"(tabs)/home"} />
-    );
+  return <Redirect href={"/(tabs)/home/home"} />;
 }
